feat : added validation for user and optimize the userController

+++ src/main/java/com/pos/hyper/controller/UserController.java
@@ -1,46 +1,67 @@
 package com.pos.hyper.controller;
 
-
 import com.pos.hyper.model.User;
 import com.pos.hyper.repository.UserRepository;
+import com.pos.hyper.validation.UserValidation;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/user")
 public class UserController {
+
     private final UserRepository userRepository;
+
+    @Autowired
+    private UserValidation userValidation;
 
     public UserController(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
     @GetMapping("")
-    List<User> getUsers() {
-        return this.userRepository.findAll();
+    public List<User> getUsers() {
+        return userRepository.findAll();
     }
+
     @GetMapping("/{id}")
-    User getUser(@PathVariable Long id) {
-        return this.userRepository.findById(id).get();
+    public User getUser(@PathVariable Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(
+                        HttpStatus.NOT_FOUND, "User not found with id: " + id));
     }
 
     @PostMapping("")
-    User addUser(@RequestBody User user) {
-        return this.userRepository.save(user);
+    public User addUser(@RequestBody User user) {
+        user = userValidation.userValidate(user);
+        return userRepository.save(user);
     }
 
     @PutMapping("/{id}")
-    User updateUser(@RequestBody User user, @PathVariable Long id) {
-        User usr = userRepository.findById(id).get();
-        usr.setName(user.getName());
-        usr.setEmail(user.getEmail());
-        usr.setPassword(user.getPassword());
-        usr.setRole(user.getRole());
-        return userRepository.save(usr);
+    public User updateUser(@RequestBody User user, @PathVariable Long id) {
+        user = userValidation.userValidate(user);
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(
+                        HttpStatus.NOT_FOUND, "User not found with id: " + id));
+
+        existingUser.setName(user.getName());
+        existingUser.setEmail(user.getEmail());
+        existingUser.setPassword(user.getPassword());
+        existingUser.setRole(user.getRole());
+
+        return userRepository.save(existingUser);
     }
+
     @DeleteMapping("/{id}")
-    void deleteUser(@PathVariable Long id) {
+    public void deleteUser(@PathVariable Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
